+++ lib/Neo4j/Record.pm
@@ -6,6 +6,7 @@ use utf8;
 #use Devel::StackTrace qw();
 use Data::Dumper qw();
 use Carp qw(croak);
+use JSON::PP;
 
 
 # sub test {
@@ -18,7 +19,26 @@ sub get {
 	my ($self, $field) = @_;
 	
 	return $self->{row}->[0] if ! defined $field;
-	return $self->{row}->[ $self->{column_keys}->key($field) ];
+	my $key = $self->{column_keys}->key($field);
+	croak "Field '$field' not present in query result" if ! defined $key;
+	return $self->{row}->[$key];
+}
+
+
+sub get_bool {
+	my ($self, $field) = @_;
+	
+	my $value = $self->get($field);
+	return $value if ! ref $value;
+	return $value if $value != JSON::PP::false;
+	return undef;
+}
+
+
+sub stats {
+	my ($self) = @_;
+	
+	return $self->{_stats};
 }